[Api] Moving fixtures to ApiBundle and small refactoring
