+++ docs/api/customers.rst
@@ -92,7 +92,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/customers/ \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Accept: application/json"
 
 Example Response
@@ -244,7 +244,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/customers/399 \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Accept: application/json"
 
 Example Response
@@ -317,7 +317,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/customers/ \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Content-Type: application/json" \
         -X POST \
         --data '
@@ -369,7 +369,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/customers/ \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Accept: application/json" \
         -X POST
 
@@ -461,7 +461,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/customers/399 \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Content-Type: application/json" \
         -X PUT \
         --data '
@@ -488,7 +488,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/customers/399 \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Accept: application/json" \
         -X PUT
 
@@ -577,7 +577,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/customers/399 \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Content-Type: application/json" \
         -X PATCH \
         --data '{"first_name": "Joe"}'
@@ -613,7 +613,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/customers/399 \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Accept: application/json" \
         -X DELETE
