+++ docs/api/products.rst
@@ -91,7 +91,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/products/ \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Accept: application/json"
 
 Example Response
@@ -182,7 +182,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/products/2 \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Accept: application/json"
 
 Example Response
@@ -301,7 +301,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/products/ \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Content-Type: application/json" \
         -X POST \
         --data '
@@ -360,7 +360,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/products/ \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Accept: application/json" \
         -X POST
 
@@ -456,7 +456,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/product/3 \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Content-Type: application/json" \
         -X PUT \
         --data '
@@ -485,7 +485,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/products/3 \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Accept: application/json" \
         -X PUT
 
@@ -569,7 +569,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/product/3 \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Content-Type: application/json" \
         -X PATCH \
         --data '
@@ -613,7 +613,7 @@ Example
 .. code-block:: bash
 
     curl http://sylius.dev/api/v1/products/3 \
-        -H "Authorization: Bearer SampleTokenMWExMWM0NzE1NmUyZDgyZDJiMjEzMmFlMjQ4MzgwMmE4ZTkxYzM0YjdlN2U2YzliNDIyMTk1ZDhlNDYxYWE4Ng" \
+        -H "Authorization: Bearer SampleToken" \
         -H "Accept: application/json" \
         -X DELETE
