+++ src/Sylius/Bundle/ApiBundle/Resources/config/app/config.yml
@@ -3,6 +3,7 @@
 
 imports:
     - { resource: "@SyliusApiBundle/Resources/config/grids/product.yml" }
+    - { resource: "@SyliusApiBundle/Resources/config/app/fixtures.yml" }
 
 sylius_api:
     resources:
