+++ src/Sylius/Bundle/ApiBundle/Resources/config/app/fixtures.yml
@@ -1,0 +1,21 @@
+# This file is part of the Sylius package.
+# (c) Paweł Jędrzejewski
+
+sylius_fixtures:
+    suites:
+        default:
+            fixtures:
+                api_client:
+                    options:
+                        custom:
+                            default_api_client:
+                                random_id: 'demo_client'
+                                secret: 'secret_demo_client'
+                                allowed_grant_types: ['password']
+                access_token:
+                    options:
+                        custom:
+                            default_access_token:
+                                client: 'demo_client'
+                                user: 'api@example.com'
+                                token: 'SampleToken'
