+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/fixtures_factories.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Paweł Jędrzejewski
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+    <services>
+        <service id="sylius.fixture.example_factory.api_client" class="Sylius\Bundle\ApiBundle\Fixture\Factory\ApiClientExampleFactory">
+            <argument type="service" id="fos_oauth_server.client_manager.default" />
+        </service>
+
+        <service id="sylius.fixture.example_factory.api_access_token" class="Sylius\Bundle\ApiBundle\Fixture\Factory\ApiAccessTokenExampleFactory">
+            <argument type="service" id="sylius.factory.api_access_token" />
+            <argument type="service" id="sylius.repository.admin_user" />
+            <argument type="service" id="sylius.repository.api_client" />
+        </service>
+    </services>
+</container>
