+++ src/Sylius/Bundle/ApiBundle/Tests/Fixture/ApiCredentialsFixtureTest.php
@@ -9,12 +9,12 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Bundle\CoreBundle\Tests\Fixture;
+namespace Sylius\Bundle\ApiBundle\Tests\Fixture;
 
 use Doctrine\Common\Persistence\ObjectManager;
 use Matthias\SymfonyConfigTest\PhpUnit\ConfigurationTestCaseTrait;
 use Sylius\Bundle\CoreBundle\Fixture\Factory\ExampleFactoryInterface;
-use Sylius\Bundle\CoreBundle\Fixture\ApiClientFixture;
+use Sylius\Bundle\ApiBundle\Fixture\ApiClientFixture;
 
 /**
  * @author Łukasz Chruściel <lukasz.chrusciel@lakion.com>
@@ -37,7 +37,7 @@ class OAuthCredentialsFixtureTest extends \PHPUnit_Framework_TestCase
     public function oauth_credentials_can_be_created_with_custom_random_id()
     {
         $this->assertConfigurationIsValid([['custom' => [[
-                'random_id' => 'totally_random',
+            'random_id' => 'totally_random',
         ]]]], 'custom.*.random_id');
     }
 
@@ -47,7 +47,7 @@ class OAuthCredentialsFixtureTest extends \PHPUnit_Framework_TestCase
     public function oauth_credentials_can_be_created_with_custom_secret()
     {
         $this->assertConfigurationIsValid([['custom' => [[
-                'secret' => 'threeCanKeepSecretIfTwoAreDead'
+            'secret' => 'threeCanKeepSecretIfTwoAreDead',
         ]]]], 'custom.*.secret');
     }
 
@@ -58,7 +58,7 @@ class OAuthCredentialsFixtureTest extends \PHPUnit_Framework_TestCase
     {
         $this->assertConfigurationIsValid([['custom' => [[
             'allowed_grant_types' => [
-                'password'
+                'password',
             ],
         ]]]], 'custom.*.allowed_grant_types');
     }
